requirements.txt
tensorflow==2.4.1
tensorflow-addons==0.12.1
tensorflow-probability==0.12.1
tensorflow-hub==0.11.0
gym==0.18.0
numpy~=1.19.5
pandas~=1.2.3
six~=1.15.0
neptune-client